actually keeping the core reader function in dl1_data_handler and have just the minimum in the pluging module might be better
